--- book/src/main/java/com/example/member/book/dto/request/RentalCreateReq.java
@@ -1,18 +1,0 @@
-package com.example.member.book.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.time.LocalDate;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class RentalCreateReq {
-    private Long userId;
-    private Long bookId;
-    private LocalDate dueDate;
-}
