--- book/src/main/java/com/example/member/book/entity/Author.java
@@ -1,29 +1,0 @@
-package com.example.member.book.entity;
-
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-@Table(name = "author")
-public class Author {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long authorId;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private String country;
-
-    @OneToMany(mappedBy = "author")
-    private List<Book> books = new ArrayList<>();
-}
