--- book/src/main/java/com/example/member/book/entity/Genre.java
@@ -1,29 +1,0 @@
-package com.example.member.book.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-@Table(name = "genre")
-public class Genre {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long genreId;
-
-    @Column(nullable = false)
-    private String genreName;
-
-    @Column
-    private String description;
-
-    @OneToMany(mappedBy = "genre")
-    private List<BookGenre> bookGenres = new ArrayList<>();
-
-}
