--- book/src/main/java/com/example/member/book/entity/Rental.java
@@ -1,42 +1,0 @@
-package com.example.member.book.entity;
-
-import com.example.member.user.domain.MemberEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDate;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Table(name = "rental")
-public class Rental {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long rentalId;
-
-    @Column(nullable = false)
-    private LocalDate rentalDate;
-
-    @Column(nullable = false)
-    private LocalDate dueDate;
-
-    @Column
-    private LocalDate returnDate;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private MemberEntity user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "book_id")
-    private Book book;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private RentalStatus status = RentalStatus.RENTED;
-}
