--- book/src/main/java/com/example/member/book/entity/RentalStatus.java
@@ -1,6 +1,0 @@
-package com.example.member.book.entity;
-
-public enum RentalStatus {
-    RENTED,
-    RETURNED
-}
