--- book/src/main/java/com/example/member/book/repository/AuthorRepository.java
@@ -1,7 +1,0 @@
-package com.example.member.book.repository;
-
-import com.example.member.book.entity.Author;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AuthorRepository extends JpaRepository<Author, Long> {
-}
