--- book/src/main/java/com/example/member/book/repository/BookGenreRepository.java
@@ -1,7 +1,0 @@
-package com.example.member.book.repository;
-
-import com.example.member.book.entity.BookGenre;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BookGenreRepository extends JpaRepository<BookGenre, Long> {
-}
