--- book/src/main/java/com/example/member/book/repository/BookRepository.java
@@ -1,7 +1,0 @@
-package com.example.member.book.repository;
-
-import com.example.member.book.entity.Book;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BookRepository extends JpaRepository<Book, Long> {
-}
