--- book/src/main/java/com/example/member/book/repository/GenreRepository.java
@@ -1,7 +1,0 @@
-package com.example.member.book.repository;
-
-import com.example.member.book.entity.Genre;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface GenreRepository extends JpaRepository<Genre, Long> {
-}
