+++ book/src/main/java/com/example/member/domain/author/entity/Author.java
@@ -1,0 +1,30 @@
+package com.example.member.domain.author.entity;
+
+
+import com.example.member.domain.book.entity.Book;
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Builder
+@Table(name = "author")
+public class Author {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long authorId;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Column(nullable = false)
+    private String country;
+
+    @OneToMany(mappedBy = "author")
+    private List<Book> books = new ArrayList<>();
+}
