+++ book/src/main/java/com/example/member/domain/author/repository/AuthorRepository.java
@@ -1,0 +1,7 @@
+package com.example.member.domain.author.repository;
+
+import com.example.member.domain.author.entity.Author;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AuthorRepository extends JpaRepository<Author, Long> {
+}
