+++ book/src/main/java/com/example/member/domain/book/repository/BookGenreRepository.java
@@ -1,0 +1,7 @@
+package com.example.member.domain.book.repository;
+
+import com.example.member.domain.book.entity.BookGenre;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BookGenreRepository extends JpaRepository<BookGenre, Long> {
+}
