+++ book/src/main/java/com/example/member/domain/book/repository/BookRepository.java
@@ -1,0 +1,7 @@
+package com.example.member.domain.book.repository;
+
+import com.example.member.domain.book.entity.Book;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BookRepository extends JpaRepository<Book, Long> {
+}
