+++ book/src/main/java/com/example/member/domain/genre/entity/Genre.java
@@ -1,0 +1,30 @@
+package com.example.member.domain.genre.entity;
+
+import com.example.member.domain.book.entity.BookGenre;
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Builder
+@Table(name = "genre")
+public class Genre {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long genreId;
+
+    @Column(nullable = false)
+    private String genreName;
+
+    @Column
+    private String description;
+
+    @OneToMany(mappedBy = "genre")
+    private List<BookGenre> bookGenres = new ArrayList<>();
+
+}
