+++ book/src/main/java/com/example/member/domain/genre/repository/GenreRepository.java
@@ -1,0 +1,7 @@
+package com.example.member.domain.genre.repository;
+
+import com.example.member.domain.genre.entity.Genre;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface GenreRepository extends JpaRepository<Genre, Long> {
+}
