+++ book/src/main/java/com/example/member/domain/home/controller/HomeController.java
@@ -1,0 +1,14 @@
+package com.example.member.domain.home.controller;
+
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class HomeController {
+    // 기본 페이지 요청 메서드
+    @GetMapping("/")
+    public String index() {
+        return "index";
+    }
+}
