+++ book/src/main/java/com/example/member/domain/rental/dto/request/RentalCreateReq.java
@@ -1,0 +1,18 @@
+package com.example.member.domain.rental.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.time.LocalDate;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class RentalCreateReq {
+    private Long userId;
+    private Long bookId;
+    private LocalDate dueDate;
+}
