+++ book/src/main/java/com/example/member/domain/rental/entity/Rental.java
@@ -1,0 +1,43 @@
+package com.example.member.domain.rental.entity;
+
+import com.example.member.domain.book.entity.Book;
+import com.example.member.domain.user.domain.MemberEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.time.LocalDate;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Table(name = "rental")
+public class Rental {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long rentalId;
+
+    @Column(nullable = false)
+    private LocalDate rentalDate;
+
+    @Column(nullable = false)
+    private LocalDate dueDate;
+
+    @Column
+    private LocalDate returnDate;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private MemberEntity user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "book_id")
+    private Book book;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private RentalStatus status = RentalStatus.RENTED;
+}
