+++ book/src/main/java/com/example/member/domain/rental/entity/RentalStatus.java
@@ -1,0 +1,6 @@
+package com.example.member.domain.rental.entity;
+
+public enum RentalStatus {
+    RENTED,
+    RETURNED
+}
