--- book/src/main/java/com/example/member/user/controller/MemberController.java
@@ -1,103 +1,0 @@
-package com.example.member.user.controller;
-
-import com.example.member.user.dto.MemberDTO;
-import com.example.member.user.service.MemberService;
-import jakarta.servlet.http.HttpSession;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@Controller
-@RequiredArgsConstructor
-public class MemberController {
-
-    private final MemberService memberService;
-
-    @GetMapping("/member/save")
-    public String saveForm() {
-        return "save";
-    }
-
-    @PostMapping("/member/save")
-    public String save(@ModelAttribute MemberDTO memberDTO) {
-        System.out.println(memberDTO);
-        memberService.save(memberDTO);
-        return "login";
-    }
-
-    @GetMapping("/member/login")
-    public String loginForm() {
-        return "login";
-    }
-
-    @PostMapping("/member/login")
-    public String login(@ModelAttribute MemberDTO memberDTO, HttpSession session) {
-        MemberDTO loginResult  = memberService.login(memberDTO);
-        if(loginResult != null) {
-            // 로그인 성공
-            session.setAttribute("loginEmail", loginResult.getMemberEmail());
-            session.setAttribute("loginId", loginResult.getId()); // userId 저장도 같이
-            return "redirect:/main";
-        } else {
-            // 로그인 실패
-            return "login";
-        }
-    }
-
-    @GetMapping("/main")
-    public String mainPage() {
-        return "user/main";
-    }
-
-    @GetMapping("/member/")
-    public String findAll( Model model ) {
-        List<MemberDTO> memberDTOList = memberService.findAll();
-//        어떠한 html로 가져갈 데이터가 있다면 model 사용
-        model.addAttribute("memberList", memberDTOList);
-        return "user/list";
-    }
-
-    @GetMapping("/member/{id}")
-    public String findById(@PathVariable Long id, Model model) {
-        MemberDTO memberDTO = memberService.findById(id);
-        model.addAttribute("member", memberDTO);
-        return "user/detail";
-    }
-
-    @GetMapping("/member/update")
-    public String updateForm(Model model, HttpSession session) {
-        String myEmail = (String) session.getAttribute("loginEmail");
-        MemberDTO memberDTO = memberService.updateForm(myEmail);
-        model.addAttribute("updateMember", memberDTO);
-        return "user/update";
-    }
-
-    @PostMapping("/member/update")
-    public String update(@ModelAttribute MemberDTO memberDTO) {
-        memberService.update(memberDTO);
-        return "redirect:/member/" + memberDTO.getId();
-    }
-
-    @GetMapping("/member/delete/{id}")
-    public String delete(@PathVariable Long id, Model model) {
-        memberService.deleteById(id);
-        return "redirect:/member/";
-    }
-
-    @GetMapping("/member/logout")
-    public String logout(HttpSession session) {
-        session.invalidate();
-        return "index";
-    }
-
-    @PostMapping("/member/email-check")
-    public @ResponseBody String emailCheck(@RequestParam String memberEmail) {
-        System.out.println("memberEmail = " + memberEmail);
-        String checkResult = memberService.emailCheck(memberEmail);
-        return checkResult;
-    }
-
-}
