--- book/src/main/java/com/example/member/user/domain/MemberEntity.java
@@ -1,46 +1,0 @@
-package com.example.member.user.domain;
-
-import com.example.member.user.dto.MemberDTO;
-import com.example.member.global.entity.BaseTimeEntity;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Setter
-@Getter
-@Table(name = "member_table")
-public class MemberEntity extends BaseTimeEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(unique = true)
-    private String memberEmail;
-
-    @Column
-    private String memberName;
-
-    @Column
-    private String memberPassword;
-
-    @Column
-    private String memberPhone;
-
-    public static MemberEntity toMemberEntity(MemberDTO memberDTO) {
-        MemberEntity memberEntity = new MemberEntity();
-        memberEntity.setMemberEmail(memberDTO.getMemberEmail());
-        memberEntity.setMemberName(memberDTO.getMemberName());
-        memberEntity.setMemberPassword(memberDTO.getMemberPassword());
-        return memberEntity;
-    }
-
-    public static MemberEntity toUpdateMemberEntity(MemberDTO memberDTO) {
-        MemberEntity memberEntity = new MemberEntity();
-        memberEntity.setId(memberDTO.getId());
-        memberEntity.setMemberEmail(memberDTO.getMemberEmail());
-        memberEntity.setMemberName(memberDTO.getMemberName());
-        memberEntity.setMemberPassword(memberDTO.getMemberPassword());
-        return memberEntity;
-    }
-}
