--- book/src/main/java/com/example/member/user/dto/MemberDTO.java
@@ -1,29 +1,0 @@
-package com.example.member.user.dto;
-
-import com.example.member.user.domain.MemberEntity;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-public class MemberDTO {
-    private Long id;
-    private String memberName;
-    private String memberEmail;
-    private String memberPassword;
-    private String memberPhone;
-
-    public static MemberDTO toMemberDTO(MemberEntity memberEntity) {
-        MemberDTO memberDTO = new MemberDTO();
-        memberDTO.setId(memberEntity.getId());
-        memberDTO.setMemberName(memberEntity.getMemberName());
-        memberDTO.setMemberEmail(memberEntity.getMemberEmail());
-        memberDTO.setMemberPassword(memberEntity.getMemberPassword());
-        memberDTO.setMemberPhone(memberEntity.getMemberPhone());
-        return memberDTO;
-    }
-}
