--- book/src/main/java/com/example/member/user/service/MemberService.java
@@ -1,93 +1,0 @@
-package com.example.member.user.service;
-
-
-import com.example.member.user.domain.MemberEntity;
-import com.example.member.user.dto.MemberDTO;
-import com.example.member.user.repository.MemberRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-
-@Service
-@RequiredArgsConstructor
-public class MemberService {
-    private final MemberRepository memberRepository;
-
-    public void save(MemberDTO memberDTO) {
-        MemberEntity memberEntity = MemberEntity.toMemberEntity(memberDTO);
-        memberRepository.save(memberEntity);
-
-    }
-
-    public MemberDTO login(MemberDTO memberDTO) {
-//        1. 회원이 입력한 이메일로 DB 조회
-//        2. DB에서 조회한 비밀번호와 사용자가 입력한 비밀번호 일치 확인
-        Optional<MemberEntity> byMemberEmail = memberRepository.findByMemberEmail(memberDTO.getMemberEmail());
-        if(byMemberEmail.isPresent()) {
-//            조회 결과가 있으면
-            MemberEntity memberEntity = byMemberEmail.get();
-            if(memberEntity.getMemberPassword().equals(memberDTO.getMemberPassword())) {
-                // 비밀번호 일치
-                // entity -> dto 변환
-                MemberDTO dto = MemberDTO.toMemberDTO(memberEntity);
-                return dto;
-            } else{
-                // 비밀번호 불일치(로그인 실패)
-                return null;
-            }
-        } else{
-
-//            조회 결과가 없으면
-            return null;
-        }
-    }
-
-    public List<MemberDTO> findAll() {
-        List<MemberEntity> memberEntityList = memberRepository.findAll();
-        List<MemberDTO> memberDTOList = new ArrayList<>();
-        for(MemberEntity memberEntity : memberEntityList) {
-            memberDTOList.add(MemberDTO.toMemberDTO(memberEntity));
-        }
-        return memberDTOList;
-    }
-
-    public MemberDTO findById(Long id) {
-        Optional<MemberEntity>  optionalMemberEntity = memberRepository.findById(id);
-        if (optionalMemberEntity.isPresent()) {
-            return MemberDTO.toMemberDTO(optionalMemberEntity.get());
-        } else {
-            return null;
-        }
-    }
-
-    public MemberDTO updateForm(String myEmail) {
-        Optional<MemberEntity> optionalMemberEntity = memberRepository.findByMemberEmail(myEmail);
-        if (optionalMemberEntity.isPresent()) {
-            return MemberDTO.toMemberDTO(optionalMemberEntity.get());
-        } else {
-            return null;
-        }
-    }
-
-    public void update(MemberDTO memberDTO) {
-        memberRepository.save(MemberEntity.toUpdateMemberEntity(memberDTO));
-    }
-
-    public void deleteById(Long id) {
-        memberRepository.deleteById(id);
-    }
-
-    public String emailCheck(String memberEmail) {
-        Optional<MemberEntity> optionalMemberEntity = memberRepository.findByMemberEmail(memberEmail);
-        if(optionalMemberEntity.isPresent()) {
-            // 조회결과 존재 -> 사용 불가
-            return null;
-        } else {
-            // 사용 가능
-            return "ok";
-        }
-    }
-}
